Add Update VisitService, add exceptions

+++ visit-service/src/main/java/com/visit/service/VisitService.java
@@ -1,6 +1,7 @@
 package com.visit.service;
 
 import com.visit.domain.Visit;
+import com.visit.domain.exception.VisitNotExistException;
 import com.visit.repository.VisitRepository;
 import org.springframework.stereotype.Service;
 
@@ -26,15 +27,18 @@ public class VisitService {
         return visitRepository.findAllByDoctorId(doctorId);
     }
 
-    public Visit findVisitById(final Long id) {
-        return visitRepository.findById(id).orElseThrow(IllegalAccessError::new);
+    public Visit findVisitById(final Long id) throws VisitNotExistException {
+        return visitRepository.findById(id).orElseThrow(VisitNotExistException::new);
     }
 
     public Visit createVisit(final Visit visit) {
         return visitRepository.save(visit);
     }
 
-    public void deleteVisit(final Visit visit) {
+    public void deleteVisit(final Visit visit) throws VisitNotExistException {
+        if(!visitRepository.existsById(visit.getId())) {
+            throw new VisitNotExistException();
+        }
         deleteVisit(visit);
     }
 }
